[nrf noup] Wi-Fi fixes and improvements
Squashed several commits including significant Wi-Fi
fixes and improvements

1. Implemented most of the Wi-Fi DiagnosticDataProvider.

Added PHY statistics:
  * multicast RX/TX
  * unicast RX/TX
  * beacons lost/TX

Signed-off-by: Marcin Kajor <marcin.kajor@nordicsemi.no>

2. wifi: map supplicant WiFi version into what Matter expects

This is needed to return proper Wi-Fi version code to the Matter
controller when querying wifinetworkdiagnostic cluster.

Signed-off-by: Marcin Kajor <marcin.kajor@nordicsemi.no>

3. Fix the 'disconnected' status appearing after scan is issued.

Restore the connection state which was set before requesting scan.
Note that the connection was not really dropped, only the status
reporting was broken.

Signed-off-by: Marcin Kajor <marcin.kajor@nordicsemi.no>

4. wifi: fixed the connection callback initialization

This fixes the misbehavior when attempting to establish a connection
with the network that is not added to the networkcommissioning cluster.

Signed-off-by: Marcin Kajor <marcin.kajor@nordicsemi.no>

5. Fixed the timeout when connecting to the WPA3 secured AP.

Set the WiFi MFP (management frame protection) as at least optional.
The current WiFi driver does not support this parameter within the
scan result, so it must the hard coded for the time being
(made it mandatory for secure associations, just in case).

Signed-off-by: Marcin Kajor <marcin.kajor@nordicsemi.no>